+++ core/src/test/java/com/rabbitmq/spark/connector/TestRoutingStrategy.java
@@ -7,7 +7,7 @@ public class TestRoutingStrategy implements ConnectorRoutingStrategy {
     private static final long serialVersionUID = 1L;
 
     @Override
-    public List<String> route(String routingKey, List<String> partitions) {
-        return partitions.subList(0, 1);
+    public List<String> route(ConnectorMessageView message, Metadata metadata) {
+        return metadata.partitions().subList(0, 1);
     }
 }